+++ id1/mathlib.h
@@ -26,7 +26,7 @@ typedef vec_t vec5_t[5];
 typedef	int	fixed4_t;
 typedef	int	fixed8_t;
 typedef	int	fixed16_t;
-typedef long long fixed44p20_t;
+typedef int64_t	fixed44p20_t;
 
 #ifndef M_PI
 #define M_PI		3.14159265358979323846	// matches value in gcc v2 math.h
